Made health card number a long

+++ app/src/main/java/com/example/seg_2105_project/Patient.java
@@ -2,15 +2,15 @@ package com.example.seg_2105_project;
 
 public class Patient extends User {
 
-    private int healthCardNumber;
+    private long healthCardNumber;
 
     public Patient() {}
-    public Patient(String firstName, String lastName, String email, String password, long phoneNumber, String address, int healthCardNumber) {
+    public Patient(String firstName, String lastName, String email, String password, long phoneNumber, String address, long healthCardNumber) {
         super(firstName, lastName, email, password, phoneNumber, address);
         this.healthCardNumber = healthCardNumber;
     }
 
-    public int getHealthCardNumber() {
+    public long getHealthCardNumber() {
         return healthCardNumber;
     }
 }
